requirements.txt
flask
pandas
numpy
scikit-learn
gunicorn
classifier
tiktoken
openai
langchain-openai
crewai
crewai_tools
load_dotenv
langchain-huggingface
PyPDF2
requests
pymongo
flask-cors
crewai-tools
weaviate
langchain-weaviate
weaviate-client
pinecone
jupyter
sentence-transformers
pinecone-text
pinecone-notebooks
fitz